[Refactor] 코드리뷰 반영 (2/2)
- 사용되지 않는 FrameLayout 제거
- Glide로 이미지 띄우기
