+++ app/src/main/AndroidManifest.xml
@@ -2,6 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="org.sopt.soptseminar_week1">
 
+    <uses-permission android:name="android.permission.INTERNET" />
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -11,7 +12,7 @@
         android:theme="@style/Theme.SoptSeminar_week1">
         <activity android:name=".view.HomeActivity" />
         <activity android:name=".view.SignUpActivity" />
-        <activity android:name=".view.UserInfoActivity"/>
+        <activity android:name=".view.UserInfoActivity" />
         <activity android:name=".view.SignInActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
