+++ app/src/main/java/org/sopt/soptseminar_week1/data/FollowingListInfo.kt
@@ -1,4 +1,4 @@
-package org.sopt.soptseminar_week1
+package org.sopt.soptseminar_week1.data
 
 data class FollowingListUserInfo(
     val userName: String
