+++ app/src/main/java/org/sopt/soptseminar_week1/view/FollowingListAdapter.kt
@@ -1,9 +1,12 @@
 package org.sopt.soptseminar_week1.view
 
+import android.content.Context
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import org.sopt.soptseminar_week1.FollowingListUserInfo
+import com.bumptech.glide.Glide
+import org.sopt.soptseminar_week1.data.FollowingListUserInfo
 import org.sopt.soptseminar_week1.databinding.ItemFollowingListBinding
 
 class FollowingListAdapter : RecyclerView.Adapter<FollowingListAdapter.FollowingUserViewHolder>() {
@@ -16,8 +19,10 @@ class FollowingListAdapter : RecyclerView.Adapter<FollowingListAdapter.Following
 
     class FollowingUserViewHolder(private val binding: ItemFollowingListBinding) :
         RecyclerView.ViewHolder(binding.root) {
-        fun onBind(followingListUserInfo: FollowingListUserInfo) {
+        fun onBind(followingListUserInfo: FollowingListUserInfo, context: Context) {
             binding.textFollowingListProfile.text = followingListUserInfo.userName
+            Glide.with(context).load(followingListUserInfo.userImageSrc).into(binding.imgFollowingListProfile)
+            Log.d("로그", "리사이클러뷰 만들어짐 ${followingListUserInfo.userImageSrc}")
         }
     }
 
@@ -30,7 +35,7 @@ class FollowingListAdapter : RecyclerView.Adapter<FollowingListAdapter.Following
     override fun getItemCount(): Int = userList.size
 
     override fun onBindViewHolder(holder: FollowingUserViewHolder, position: Int) {
-        holder.onBind(userList[position])
+        holder.onBind(userList[position], holder.itemView.getContext())
     }
 
 }
