+++ app/src/main/res/layout/item_following_list.xml
@@ -18,6 +18,8 @@
         android:id="@+id/text_following_list_profile"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:paddingStart="10dp"
+        android:paddingEnd="10dp"
         android:textColor="@color/black"
         android:textSize="18sp"
         app:layout_constraintBottom_toBottomOf="parent"
